Recent version of JQuery catches this syntax error ',editablefields-processed'
Recent version of JQuery catches this syntax error ',editablefields-processed'
and generates "Uncaught Syntax error, unrecognized expression"
which leads to "Error, unable to make update" messagebox in browser
when entering student grades.
There is a new version of editablefields available, but we use a customised version.
I have ported our changes to this new editablefields module, but to my best knowledge,
this has not been tested, so I do not want to commit. I have therefore committed
this minimal fix.